Move XMLParser into its own file
